+++ dsp/src/tone.rs
@@ -11,23 +11,65 @@ pub struct Attributes {
 
 #[derive(Debug)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct Tone {
+pub struct Tone2 {
     sample_rate: f32,
+    pub tone_1: Tone,
+    pub tone_2: Tone,
+}
+
+#[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub struct Tone {
     lpf: LinkwitzRileyFilter,
     hpf: LinkwitzRileyFilter,
 }
 
-impl Tone {
+impl Tone2 {
     /// # Panics
     ///
     /// Sample rate must be higher than 500, otherwise the filter becomes
     /// unstable and the initialization will panic.
     #[must_use]
-    pub fn new(sample_rate: u32) -> Self {
+    pub fn new(sample_rate: f32) -> Self {
         Self {
-            sample_rate: sample_rate as f32,
-            lpf: LinkwitzRileyFilter::new(sample_rate as f32),
-            hpf: LinkwitzRileyFilter::new(sample_rate as f32),
+            sample_rate,
+            tone_1: Tone::new(sample_rate),
+            tone_2: Tone::new(sample_rate),
+        }
+    }
+
+    pub fn set_attributes(&mut self, attributes: Attributes) {
+        let a = 13.73;
+        if attributes.tone < 0.4 {
+            self.tone_1.hpf.set_frequency(0.0);
+            self.tone_2.hpf.set_frequency(0.0);
+            let phase = attributes.tone / 0.4;
+            let voct = phase * 10.645;
+            let cutoff = a * libm::powf(2.0, voct);
+            self.tone_1.lpf.set_frequency(cutoff);
+            self.tone_2.lpf.set_frequency(cutoff);
+        } else if attributes.tone < 0.6 {
+            self.tone_1.lpf.set_frequency(self.sample_rate * 0.48);
+            self.tone_2.lpf.set_frequency(self.sample_rate * 0.48);
+            self.tone_1.hpf.set_frequency(0.0);
+            self.tone_2.hpf.set_frequency(0.0);
+        } else {
+            self.tone_1.lpf.set_frequency(self.sample_rate * 0.48);
+            self.tone_2.lpf.set_frequency(self.sample_rate * 0.48);
+            let phase = (attributes.tone - 0.6) / 0.4;
+            let voct = phase * 10.0;
+            let cutoff = a * libm::powf(2.0, voct);
+            self.tone_1.hpf.set_frequency(cutoff);
+            self.tone_2.hpf.set_frequency(cutoff);
+        }
+    }
+}
+
+impl Tone {
+    fn new(sample_rate: f32) -> Self {
+        Self {
+            lpf: LinkwitzRileyFilter::new(sample_rate),
+            hpf: LinkwitzRileyFilter::new(sample_rate),
         }
     }
 
@@ -38,26 +80,6 @@ impl Tone {
     pub fn process(&mut self, buffer: &mut [f32]) {
         for x in buffer.iter_mut() {
             *x = self.tick(*x);
-        }
-    }
-
-    pub fn set_attributes(&mut self, attributes: Attributes) {
-        let a = 13.73;
-        if attributes.tone < 0.4 {
-            self.hpf.set_frequency(0.0);
-            let phase = attributes.tone / 0.4;
-            let voct = phase * 10.645;
-            let cutoff = a * libm::powf(2.0, voct);
-            self.lpf.set_frequency(cutoff);
-        } else if attributes.tone < 0.6 {
-            self.lpf.set_frequency(self.sample_rate * 0.48);
-            self.hpf.set_frequency(0.0);
-        } else {
-            self.lpf.set_frequency(self.sample_rate * 0.48);
-            let phase = (attributes.tone - 0.6) / 0.4;
-            let voct = phase * 10.0;
-            let cutoff = a * libm::powf(2.0, voct);
-            self.hpf.set_frequency(cutoff);
         }
     }
 }